+++ includes/ConnectDB_mariadb.php
@@ -1,17 +1,18 @@
 <?php
-/* Database abstraction for MariaDB (based on mysqli) */
+
+/* Database abstraction for MariaDB (based on the php mysqli extension) */
 
 define ('LIKE', 'LIKE');
 
-if (!isset($MySQLPort)) {
-	$MySQLPort = 3306;
+if (!isset($DBPort)) {
+	$DBPort = 3306;
 }
 global $db;	// Make sure it IS global, regardless of our context
 
 // since php 8.1, failures to connect will throw an exception, preventing our own error handling. Reset that
 mysqli_report(MYSQLI_REPORT_ERROR);
 
-$db = mysqli_connect($Host, $DBUser, $DBPassword, $_SESSION['DatabaseName'], $MySQLPort);
+$db = mysqli_connect($Host, $DBUser, $DBPassword, $_SESSION['DatabaseName'], $DBPort);
 
 /* check connection */
 if (mysqli_connect_errno()) {
@@ -32,18 +33,19 @@ if (!$db) {
 mysqli_set_charset($db, 'utf8');
 
 /* Update to allow RecurringSalesOrdersProcess.php to run via cron */
+/// @todo test if this is in fact necessary, as RecurringSalesOrdersProcess.php also sets $_SESSION['DatabaseName']
 if (isset($DatabaseName)) {
 	if (!mysqli_select_db($db, $DatabaseName)) {
 		echo '<br />' . __('The company name entered does not correspond to a database on the database server specified in the config.php configuration file. Try logging in with a different company name');
 		echo '<br /><a href="' . $RootPath . '/index.php">' . __('Back to login page') . '</a>';
-		unset ($DatabaseName);
+		//unset ($DatabaseName);
 		exit();
 	}
 } else {
 	if (!mysqli_select_db($db, $_SESSION['DatabaseName'])) {
 		echo '<br />' . __('The company name entered does not correspond to a database on the database server specified in the config.php configuration file. Try logging in with a different company name');
 		echo '<br /><a href="' . $RootPath . '/index.php">' . __('Back to login page') . '</a>';
-		unset ($_SESSION['DatabaseName']);
+		//unset ($_SESSION['DatabaseName']);
 		exit();
 	}
 }